switch to long

+++ src/cuda_dpm.h
@@ -16,7 +16,7 @@
 using namespace std;
 
 __device__ __constant__ int d_NDIM = 2;
-__device__ __constant__ int d_numVertices;
+__device__ __constant__ long d_numVertices;
 __device__ __constant__ double d_rho0;
 __device__ __constant__ double d_L[2];
 __device__ __constant__ double d_kc;
